Fix `meta.selector.css` to not extend outside the style block if embedded.
Affects embedded CSS (e.g. in HTML or in a Markdown variant supporting code blocks with syntax highlighting) only:

While typing a selector, the pattern for `meta.selector.css` would swallow up everything from the cursor until the end of the document - extending beyond the end of the embedded CSS block and thus destroying syntax highlighting for the rest of the document.