Add rebeccapurple to the list of colors
This commit updates both the CSS grammar and the Insert Color command
with rebeccapurple, a relatively new (circa 2014) CSS color.

This commit adds rebeccapurple as a "deprecated" color to make this
change as minimal as possible and because it does not work in older
browsers (those released before June 2014).

The story behind this particular color can be read here:
https://meyerweb.com/eric/thoughts/2014/06/19/rebeccapurple/